fixed collision system for python 3 compatibility, difference in integer division handling

+++ src/players.py
@@ -50,78 +50,78 @@ class KeyControl(object):
 
 
 class Inmate(object):
-    def __init__(self,parent):
-        self.parent = parent
-        self.alive = True
-        self.x = -1
-        self.y = -1
-        self.heading = 0
-        self.moving = False
-        self.speed = 4
-        self.speed_d = 3
-        self.control = KeyControl(resources.controlmap)
-    
-    def UpdateMotion(self):
-        
-        # dead?
-        if not self.alive:
-        	return
-        
-        tsize = self.parent.tiledlayers.tilesize
-        init_tile = self.parent.tiledlayers.map_size[0]*int(self.y/tsize)+int(self.x/tsize)
-        
-        # Check for player input (motion)
-        vx = self.control.r-self.control.l
-        vy = self.control.d-self.control.u
-        
-        if abs(vx) > 0 or abs(vy) > 0:
-        	self.moving = True
-        else:
-        	self.moving = False
-        
-        if abs(vx) > 0 and abs(vy) > 0:
-            self.x = self.x + self.speed_d*vx
-            self.y = self.y + self.speed_d*vy
-        else:
-            self.x = self.x + self.speed*vx
-            self.y = self.y + self.speed*vy
-        
-        # Set heading info
-        if vx > 0:
-            self.heading = 2
-        elif vx < 0:
-            self.heading = 3
-        elif vy > 0:
-        	self.heading = 1
-        elif vy < 0:
-        	self.heading = 0
-        
-        # Handle world boundaries
-        if self.x < (tsize/2):
-        	self.x = (tsize/2)
-        elif self.x > self.parent.tiledlayers.map_size[0]*tsize-(tsize/2):
-        	self.x = self.parent.tiledlayers.map_size[0]*tsize-(tsize/2)
-        if self.y < (tsize/2):
-        	self.y = (tsize/2)
-        elif self.y > self.parent.tiledlayers.map_size[1]*tsize-(tsize/2):
-        	self.y = self.parent.tiledlayers.map_size[1]*tsize-(tsize/2)
-        
-        # Handle all collisions off solid objects
-        playersize_x = 24
-        playersize_y = 24
-        (self.x,self.y) = self.parent.tiledlayers.HandleObjectWallCollision((self.x,self.y),(playersize_x,playersize_y),(vx,vy))
-        
-        # Check if need to inform tilemap object layer of updates
-        final_tile = self.parent.tiledlayers.map_size[0]*int(self.y/tsize)+int(self.x/tsize)
-        if not init_tile == final_tile:
-            self.parent.tiledlayers.UpdateObj(init_tile,final_tile,0)
-    
-    def Draw(self,screen):
-    	ts = self.parent.tiledlayers.tilesize
-        tilecoords = resources.charsprites_coords[0]
-        imw = 32
-        imh = 48
-        boxw = 24
-        boxh = 24
-        pygame.draw.rect(screen, (255,0,0), pygame.Rect((self.x-(boxw/2)-self.parent.camera.x+self.parent.camera.w_view/2,self.y-(boxh/2)-self.parent.camera.y+self.parent.camera.h_view/2),(boxw,boxh)))
-        screen.blit(resources.charsprites, (self.x-(imw/2)-self.parent.camera.x+self.parent.camera.w_view/2,self.y-(imh/2)-self.parent.camera.y+self.parent.camera.h_view/2-20), area=tilecoords)
+	def __init__(self,parent):
+		self.parent = parent
+		self.alive = True
+		self.x = -1
+		self.y = -1
+		self.heading = 0
+		self.moving = False
+		self.speed = 4
+		self.speed_d = 3
+		self.control = KeyControl(resources.controlmap)
+	
+	def UpdateMotion(self):
+	
+		# dead?
+		if not self.alive:
+			return
+		
+		tsize = self.parent.tiledlayers.tilesize
+		init_tile = self.parent.tiledlayers.map_size[0]*int(self.y/tsize)+int(self.x/tsize)
+		
+		# Check for player input (motion)
+		vx = self.control.r-self.control.l
+		vy = self.control.d-self.control.u
+		
+		if abs(vx) > 0 or abs(vy) > 0:
+			self.moving = True
+		else:
+			self.moving = False
+		
+		if abs(vx) > 0 and abs(vy) > 0:
+			self.x = self.x + self.speed_d*vx
+			self.y = self.y + self.speed_d*vy
+		else:
+			self.x = self.x + self.speed*vx
+			self.y = self.y + self.speed*vy
+		
+		# Set heading info
+		if vx > 0:
+			self.heading = 2
+		elif vx < 0:
+			self.heading = 3
+		elif vy > 0:
+			self.heading = 1
+		elif vy < 0:
+			self.heading = 0
+		
+		# Handle world boundaries
+		if self.x < (tsize/2):
+			self.x = int(tsize/2)
+		elif self.x > self.parent.tiledlayers.map_size[0]*tsize-int(tsize/2):
+			self.x = self.parent.tiledlayers.map_size[0]*tsize-int(tsize/2)
+		if self.y < (tsize/2):
+			self.y = int(tsize/2)
+		elif self.y > self.parent.tiledlayers.map_size[1]*tsize-int(tsize/2):
+			self.y = self.parent.tiledlayers.map_size[1]*tsize-int(tsize/2)
+		
+		# Handle all collisions off solid objects
+		playersize_x = 24
+		playersize_y = 24
+		(self.x,self.y) = self.parent.tiledlayers.HandleObjectWallCollision((self.x,self.y),(playersize_x,playersize_y),(vx,vy))
+		
+		# Check if need to inform tilemap object layer of updates
+		final_tile = self.parent.tiledlayers.map_size[0]*int(self.y/tsize)+int(self.x/tsize)
+		if not init_tile == final_tile:
+			self.parent.tiledlayers.UpdateObj(init_tile,final_tile,0)
+	
+	def Draw(self,screen):
+		ts = self.parent.tiledlayers.tilesize
+		tilecoords = resources.charsprites_coords[0]
+		imw = 32
+		imh = 48
+		boxw = 24
+		boxh = 24
+		pygame.draw.rect(screen, (255,0,0), pygame.Rect((self.x-(boxw/2)-self.parent.camera.x+int(self.parent.camera.w_view/2),self.y-int(boxh/2)-self.parent.camera.y+int(self.parent.camera.h_view/2)),(boxw,boxh)))
+		screen.blit(resources.charsprites, (self.x-int(imw/2)-self.parent.camera.x+int(self.parent.camera.w_view/2),self.y-int(imh/2)-self.parent.camera.y+int(self.parent.camera.h_view/2)-20), area=tilecoords)
